Remove hidden mod for .path
